filter function as callback, improve docs
