+++ lib/ex_sieve.ex
@@ -3,7 +3,7 @@ defmodule ExSieve do
   ExSieve is a object query translator to Ecto.Query.
   """
 
-  alias ExSieve.{Builder, Config, Node}
+  alias ExSieve.Config
 
   @doc """
   ExSieve is meant to be `use`d by a Ecto.Repo.
@@ -26,10 +26,12 @@ defmodule ExSieve do
 
   defmacro __using__(opts) do
     quote do
+      @behaviour ExSieve
+
       @ex_sieve_defaults unquote(opts)
 
       def filter(queryable, params, options \\ %{}) do
-        ExSieve.filter(queryable, params, Config.new(@ex_sieve_defaults, options))
+        ExSieve.Filter.filter(queryable, params, Config.new(@ex_sieve_defaults, options))
       end
     end
   end
@@ -37,28 +39,32 @@ defmodule ExSieve do
   @typep error :: :invalid_query | :attribute_not_found | :predicate_not_found | :direction_not_found | :value_is_empty
   @type result :: Ecto.Query.t() | {:error, error}
 
-  @spec filter(Ecto.Queryable.t(), %{(binary | atom) => term}, Config.t()) :: result
-  def filter(queryable, params, %Config{} = config) do
-    case extract_schema(queryable) do
-      {:ok, schema} ->
-        params
-        |> Node.call(schema, config)
-        |> result(queryable)
+  @doc """
+  Filters the given query based on params.
 
-      err ->
-        err
-    end
-  end
+  Returns the query with the added filters or an `{}:error, error}` tuple.
 
-  defp result({:error, reason}, _queryable), do: {:error, reason}
-  defp result({:ok, groupings, sorts}, queryable), do: Builder.call(queryable, groupings, sorts)
+  In order to avoid duplicated joins being sent to database only named bindings should be used
+  and the binding name should correspond to the related table one.
 
-  defp extract_schema(%Ecto.Query{from: %{source: {_, module}}}), do: extract_schema(module)
+  ## Examples
 
-  defp extract_schema(schema) when is_atom(schema) do
-    cond do
-      function_exported?(schema, :__schema__, 1) -> {:ok, schema}
-      true -> {:error, :invalid_query}
-    end
-  end
+      Repo.filter(User, %{"name_cont" => "foo"})
+
+      Repo.filter(from(u in User), %{"name_cont" => "foo"})
+
+      Repo.filter(from(u in User), %{"name_cont" => "foo"})
+
+      User
+      |> join(:inner, [u], p in assoc(u, :posts), as: :posts)
+      |> preload(:posts)
+      |> Repo.filter(%{"name_cont" => "foo"})
+
+      # WARNING: this will result in a duplicated join
+      User
+      |> join(:inner, [u], p in assoc(u, :posts), as: :posts_dup)
+      |> Repo.filter(%{"posts_title_cont" => "foo"})
+
+  """
+  @callback filter(Ecto.Queryable.t(), params :: %{(binary | atom) => term}, options :: %{atom => term}) :: result()
 end
