+++ test/ex_sieve_test.exs
@@ -13,7 +13,7 @@ defmodule ExSieveTest do
     test "return ordered by id and body", %{config: config} do
       [%{body: body} | _] = insert_pair(:post)
 
-      ids = Comment |> ExSieve.filter(%{"post_body_in" => [body], "s" => "post_id desc"}, config) |> ids
+      ids = Comment |> ExSieve.Filter.filter(%{"post_body_in" => [body], "s" => "post_id desc"}, config) |> ids
 
       ecto_ids =
         Comment
@@ -30,7 +30,7 @@ defmodule ExSieveTest do
 
       [%{body: body} | _] = insert_pair(:post)
 
-      ids = Comment |> ExSieve.filter(%{"post_body" => [body], "s" => "post_id desc"}, config) |> ids
+      ids = Comment |> ExSieve.Filter.filter(%{"post_body" => [body], "s" => "post_id desc"}, config) |> ids
       ecto_ids = Comment |> order_by([c], desc: :post_id) |> ids
       assert ids == ecto_ids
     end
@@ -40,7 +40,7 @@ defmodule ExSieveTest do
 
       [%{body: body} | _] = insert_pair(:post)
 
-      ids = Comment |> ExSieve.filter(%{"post_body_in" => [body], "s" => "posts desc"}, config) |> ids
+      ids = Comment |> ExSieve.Filter.filter(%{"post_body_in" => [body], "s" => "posts desc"}, config) |> ids
 
       ecto_ids =
         Comment
@@ -53,7 +53,7 @@ defmodule ExSieveTest do
 
     test "return users with custom type field", %{config: config} do
       insert_pair(:user)
-      ids = User |> ExSieve.filter(%{"cash_lteq" => %Money{amount: 1000}}, config) |> ids()
+      ids = User |> ExSieve.Filter.filter(%{"cash_lteq" => %Money{amount: 1000}}, config) |> ids()
       assert ids == ids(User)
     end
   end
